feat: 42 | Replace the main screens
