+++ packages/web/src/components/ChatBot/ChatBotMessageAnswer.vue
@@ -47,6 +47,7 @@ const currentFile = computed(() => {
   &__body {
     font-size: 16px;
     line-height: 1.5;
+    word-break: break-all;
   }
 
   &--notification {
