+++ packages/web/src/components/ChatBot/ChatBotMessageNotification.vue
@@ -161,6 +161,10 @@ function resetForm() {
     line-height: 1.5;
     padding-right: 48px;
     padding-bottom: 16px;
+    word-break: break-all;
+    ul, ol{
+      padding: 0 0 0 20px;
+    }
   }
 
   &__input_edit {
