+++ packages/web/src/components/ChatBot/ChatBotMessageQuestion.vue
@@ -86,6 +86,10 @@ function onClickRollbackQuestion() {
     line-height: 1.5;
     padding-right: 48px;
     padding-bottom: 16px;
+    word-break: break-all;
+    ul, ol{
+      padding: 0 0 0 20px;
+    }
   }
 
   &__actions {
